% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_wattbike_to_sb.R
\name{push_wattbike_to_sb}
\alias{push_wattbike_to_sb}
\title{Pull wattbike time trial data from the shared drive and push to Smartabase}
\usage{
push_wattbike_to_sb(
  user = Sys.info()[["user"]],
  start_folder = NA,
  end_folder = NA,
  test_duration_secs = 180,
  sb_user_id
)
}
\arguments{
\item{start_folder}{string: path to the folder containing the files to upload. No trailing slash.}

\item{end_folder}{string: path to the destination folder to relocate the files following the upload. No trailing slash.}

\item{test_duration_secs}{numeric: duration of the time trial in seconds.}

\item{sb_user_id}{numeric: uploader's Smartabase ID. Contact Smartabase administrator to obtain.}
}
\value{
Pushes data to Smartabase.
}
\description{
Pull wattbike time trial data from the shared drive and push to Smartabase
}
